put unix socket reconnections in linux only source file, stub out other OS's
