+++ recon_windows.go
@@ -1,0 +1,15 @@
+package hbot
+
+import (
+	"net"
+	"fmt"
+	"bitbucket.org/madmo/sendfd"
+)
+
+func (irc *IrcCon) StartUnixListener() { }
+
+// Attempt to hijack session previously running bot
+func (irc *IrcCon) HijackSession() bool {
+	return false
+}
+
